Add files via upload
PS2 Problem 1 bisection. The two functions are saved in ffunction and fffunction.  I wanted to pass the function into the function bisec, but this was not possible, so I just wrote two identical functions bisec and bisecc using ffunction and fffunction. I found 2 roots for the first function (-1.6xxx and 0.249) and 1 root for the second function (1).