Add files via upload
My version of PS1 P3. (Originally just due to the matter of fun ;) ). I have done no restrictions on the number of movings in the hood. However, I revieve an error in one for loop  (it is also marked) and it is only there if there are less than 3 movings in the hood in one iteration step. I just cannot see why. Matlab looks for the third entry in Mover, but I dont know why, since it only goes until length(Mover), which is 2...